Minor changes!
After a lot of trial and error with trying to convert json to sql, the attribute error with the pandas list looking for 'items' instead of 'entries' is still here. Thus instead of converting straight from json to sql, try creating graphs from json, and then use the specific keys from the plots and convert that into sql instead!