Update License
Added some icons that require attribution to the icon artists in order to use them. This made is wierd for me to keep this as a GNU 3.0 license since the art in this software now requires attribution, I'm not too fimiliar with the details of licenses so I just changed this to an MIT license to keep it simple for the future, since this repo is private for now I don't see the harm in changing the license since no one other then me has made changes to this repo yet!